refactor(nms): Move config out of fbc_js_core
