+++ nms/config/staticDist.js
