--- nms/fbc_js_core/webpack_config/paths.js
@@ -1,34 +1,0 @@
-/**
- * Copyright 2020 The Magma Authors.
- *
- * This source code is licensed under the BSD-style license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- * @flow strict-local
- * @format
- */
-
-'use strict';
-
-const fs = require('fs');
-const path = require('path');
-
-const appDirectory = fs.realpathSync(process.cwd());
-
-const resolveApp = (relativePath: string) =>
-  path.resolve(appDirectory, relativePath);
-
-module.exports = {
-  appIndexJs: resolveApp('app/main.js'),
-  loginJs: resolveApp('app/login.js'),
-  appSrc: resolveApp('app'),
-  distPath: resolveApp('static/dist'),
-  packagesDir: resolveApp('../../fbcnms-packages'),
-  resolveApp,
-};
