+++ nms/server/auth/expressOnboarding.js
@@ -18,7 +18,7 @@ import type {ExpressResponse, NextFunction} from 'express';
 import type {FBCNMSRequest} from './access';
 
 import express from 'express';
-import staticDist from '../../fbc_js_core/webpack_config/staticDist';
+import staticDist from '../../config/staticDist';
 import {AccessRoles} from '../../shared/roles';
 import {Organization, User} from '../../fbc_js_core/sequelize_models';
 import {getPropsToUpdate} from './util';
